+++ types.go
@@ -6,9 +6,9 @@ type ContainerConfig struct {
 	Hostname        string
 	Domainname      string
 	User            string
-	Memory          int
-	MemorySwap      int
-	CpuShares       int
+	Memory          int64
+	MemorySwap      int64
+	CpuShares       int64
 	Cpuset          string
 	AttachStdin     bool
 	AttachStdout    bool
@@ -51,12 +51,12 @@ type LogOptions struct {
 	Stdout     bool
 	Stderr     bool
 	Timestamps bool
-	Tail       int
+	Tail       int64
 }
 
 type RestartPolicy struct {
 	Name              string
-	MaximumRetryCount int
+	MaximumRetryCount int64
 }
 
 type PortBinding struct {
@@ -107,18 +107,18 @@ type Container struct {
 	Names      []string
 	Image      string
 	Command    string
-	Created    int
+	Created    int64
 	Status     string
 	Ports      []Port
-	SizeRw     int
-	SizeRootFs int
+	SizeRw     int64
+	SizeRootFs int64
 }
 
 type Event struct {
 	Id     string
 	Status string
 	From   string
-	Time   int
+	Time   int64
 }
 
 type Version struct {
@@ -142,13 +142,13 @@ type Image struct {
 }
 
 type Info struct {
-	Containers      int
+	Containers      int64
 	Driver          string
 	DriverStatus    [][]string
 	ExecutionDriver string
-	Images          int
+	Images          int64
 	KernelVersion   string
 	OperatingSystem string
-	NCPU            int
+	NCPU            int64
 	MemTotal        int64
 }
